Fix the path of the generated bootstrap.php.cache when using symbolic links
In SensioDistributionBundle, build_bootstrap.php uses its own path to determine
where it should genereate bootstrap.path.cache:
__DIR__.'/../../../../../../../'.'app/bootstrap.php.cache'

If your "vendor" path is just a symlink (this is the case if you use capifony
for example), this will generate bootstrap.php.cache in the wrong path.

This fix adds the root path as argument to build_bootstrap (based on the path of
bin/vendors).