Added support for my weird network config
On my dev box, I use a slightly eccentric network configuration with a second loopback, in order to make DNSMasq behave the way I want.  I can't remember all the details of why it is, but in my weird network config, the client IP doesn't show up as 127.0.0.1, even when I'm accessing the server from itself.  This fixes it for me.