Unique front controller
This PR aims at improving the front controller.

Creation of new environment:

**Before**: creation of a new front controller by duplicating `app.php`.
To access it in the browser, replace the index file by te new one (eg.
`app_dev.php/`).

**After**: addition of a line in the front controller `app.php`.
To access it in the browser, prefix the URL (eg. `app.php/dev/`).

PR #560 is not really related, but proves that a change would be
welcome.

If this solution seems acceptable, then I suppose a re-work should
be done on `.htaccess`.