Enabled translator by default
I found that when newcomers want to use translation, they struggle with the fact that it isn't enable be default. See: http://stackoverflow.com/questions/19147337/symfony-translations-not-working.
This is mainly because there's no error whatsoever.

Enabling it by default might make it easier. On the other hand it might impact performances, but I guess by the time you come to performance issues you're no longer a newcomer :) .